+++ compiler/src/main/kotlin/com/squareup/metro/extensions/developmentapp/DevelopmentAppComponentFir.kt
@@ -10,6 +10,7 @@ import com.squareup.metro.extensions.fir.hasAnnotation
 import com.squareup.metro.extensions.squareMetroExtensionsConfig
 import dev.zacsweers.metro.compiler.MetroOptions
 import dev.zacsweers.metro.compiler.api.fir.MetroFirDeclarationGenerationExtension
+import dev.zacsweers.metro.compiler.compat.CompatContext
 import org.jetbrains.kotlin.descriptors.ClassKind
 import org.jetbrains.kotlin.descriptors.Modality
 import org.jetbrains.kotlin.descriptors.Visibilities
@@ -928,6 +929,7 @@ public class DevelopmentAppComponentFir(session: FirSession) :
     override fun create(
       session: FirSession,
       options: MetroOptions,
+      compatContext: CompatContext,
     ): MetroFirDeclarationGenerationExtension = DevelopmentAppComponentFir(session)
   }
 }