+++ compiler/src/main/kotlin/com/squareup/metro/extensions/featureflag/ContributesFeatureFlagFir.kt
@@ -9,6 +9,7 @@ import com.squareup.metro.extensions.fir.buildClassExpression
 import com.squareup.metro.extensions.fir.hasAnnotation
 import dev.zacsweers.metro.compiler.MetroOptions
 import dev.zacsweers.metro.compiler.api.fir.MetroFirDeclarationGenerationExtension
+import dev.zacsweers.metro.compiler.compat.CompatContext
 import org.jetbrains.kotlin.descriptors.ClassKind
 import org.jetbrains.kotlin.descriptors.Modality
 import org.jetbrains.kotlin.descriptors.Visibilities
@@ -290,6 +291,7 @@ public class ContributesFeatureFlagFir(session: FirSession) :
     override fun create(
       session: FirSession,
       options: MetroOptions,
+      compatContext: CompatContext,
     ): MetroFirDeclarationGenerationExtension = ContributesFeatureFlagFir(session)
   }
 }