+++ compiler/src/main/kotlin/com/squareup/metro/extensions/robot/ContributesRobotMetroExtension.kt
@@ -4,6 +4,8 @@ import com.fueledbycaffeine.autoservice.AutoService
 import com.squareup.metro.extensions.fir.extractScopeClassId
 import dev.zacsweers.metro.compiler.MetroOptions
 import dev.zacsweers.metro.compiler.api.fir.MetroContributionExtension
+import dev.zacsweers.metro.compiler.compat.CompatContext
+import dev.zacsweers.metro.compiler.fir.MetroFirTypeResolver
 import org.jetbrains.kotlin.fir.FirSession
 import org.jetbrains.kotlin.fir.extensions.FirDeclarationPredicateRegistrar
 import org.jetbrains.kotlin.fir.extensions.predicateBasedProvider
@@ -40,7 +42,8 @@ public class ContributesRobotMetroExtension(private val session: FirSession) :
   }
 
   override fun getContributions(
-    scopeClassId: ClassId
+    scopeClassId: ClassId,
+    typeResolverFactory: MetroFirTypeResolver.Factory,
   ): List<MetroContributionExtension.Contribution> {
     return annotatedClasses.mapNotNull { parentSymbol ->
       val annotationScopeClassId =
@@ -76,7 +79,11 @@ public class ContributesRobotMetroExtension(private val session: FirSession) :
 
   @AutoService(MetroContributionExtension.Factory::class)
   public class Factory : MetroContributionExtension.Factory {
-    override fun create(session: FirSession, options: MetroOptions): MetroContributionExtension {
+    override fun create(
+      session: FirSession,
+      options: MetroOptions,
+      compatContext: CompatContext,
+    ): MetroContributionExtension {
       return ContributesRobotMetroExtension(session)
     }
   }